chore: Move AWS Credentials to their own file, do not pass role from external account, create dedicated Connection helper module, other misc PR cleanup
