+++ spec/googleauth/external_account_spec.rb
@@ -1,4 +1,4 @@
-# Copyright 2015 Google, Inc.
+# Copyright 2022 Google, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
